pyproject.toml
[project]
name = "nlp-linkedin-offers"
version = "0.1.0"
description = "Linkedin NLP tasks"
authors = [
    {name = "Gabriel",email = "gabriel.calvente@epita.fr"}
]
readme = "README.md"
requires-python = ">=3.12"

[tool.poetry]
name = "nlp-linkedin-offers"
version = "0.1.0"
description = "Linkedin NLP tasks"
authors = ["Gabriel <gabriel.calvente@epita.fr>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
pandas = ">=2.2.3,<3.0.0"
nltk = ">=3.9.1,<4.0.0"
tokenizers = "^0.21.1"
huggingface-hub = "^0.30.2"
mlflow = "^2.21.3"
scikit-learn = "^1.6.1"
tensorflow = "^2.19.0"
transformers = "^4.51.3"
wordcloud = "^1.9.4"
textstat = "^0.7.5"
sentence-transformers = "^4.1.0"
tf-keras = "^2.19.0"
rouge-score = "^0.1.2"
python-dotenv = "^1.1.0"
aiohttp = "^3.11.18"
evaluate = "^0.4.3"
openai = "^1.78.0"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.11.0,<0.12.0"
pyright = ">=1.1.396,<2.0.0"
mkdocs = ">=1.6.1,<2.0.0"
kagglehub = "^0.3.11"
pandas = "^2.2.3"
ipykernel = "^6.29.5"
matplotlib = "^3.10.1"
seaborn = "^0.13.2"
nbqa = "^1.9.1"
black = "^25.1.0"
isort = "^6.0.1"
torchvision = ">=0.15.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"